Made FormsMixin compatible with Dougrain v0.5.
FormsMixin.set_form and FormsMixin.delete_form use Dougrain's mutator decorator, which changed its signature in v0.5. mutator is now a function that accepts a list of caches to invalidate as optional arguments and returns the actual decorator.

FormsMixin does not invalidate any caches, so it doesn't really need mutator, but I've kept the decorator in place as an annotation on those methods.